skills/auto-deploy-bot/requirements.txt
anthropic>=0.25.0
vercel>=33.0.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
dnspython>=2.4.0
boto3>=1.26.0
pytest>=7.0.0
pytest-cov>=4.0.0
click>=8.0.0
pyyaml>=6.0
colorama>=0.4.6
